Refactors world ID to use integer type
Changes the world ID from ulong to int to improve data management and consistency.

Updates relevant classes and methods to reflect this change, including:
- MapManager
- TerrainGenerator
- PlayerManager
- WorldContainer
- WorldRender
- SignalBlockChanged
- ECSBridge

This refactor simplifies world identification and enhances overall performance.

+++ src/ModLoader/handler/IMapManager.cs
@@ -3,5 +3,5 @@
 namespace ModLoader.handler;
 
 public interface IMapManager {
-    public void RegisterGenerator<T>(ulong worldId) where T : IWorldGenerator;
+    public void RegisterGenerator<T>(int worldId) where T : IWorldGenerator;
 }
